Refactor SelectorSala
