+++ DREAM Lab Frontend/src/pages/SelectorSala/assets/images/group.png
