+++ DREAM Lab Frontend/src/pages/SelectorSala/assets/images/onePerson.png
