+++ DREAM Lab Frontend/src/pages/SelectorSala/components/Slider/Slider.css
